VC warning fixes
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk/ac@4689 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ source/src/protocol.cpp
@@ -248,7 +248,7 @@ void filterlang(char *d, const char *s)
 
 void trimtrailingwhitespace(char *s)
 {
-    for(int n = strlen(s) - 1; n >= 0 && isspace(s[n]); n--)
+    for(int n = (int)strlen(s) - 1; n >= 0 && isspace(s[n]); n--)
         s[n] = '\0';
 }
